검색한 영화 코드 정리
