+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "cross-env WEBPACK_ENV=development webpack-dev-server --open",
     "start": "cross-env WEBPACK_ENV=staging webpack-dev-server --open",
+    "build-dev": "cross-env WEBPACK_ENV=development webpack",
     "build": "cross-env WEBPACK_ENV=production webpack",
     "lint": "eslint src --ext .js,.jsx,.ts,.tsx --fix",
     "test": "jest --coverage"
